+++ R/stat_contour_fill.R
@@ -40,6 +40,7 @@ stat_contour_fill <- function(mapping = NULL, data = NULL,
 StatContourFill <- ggplot2::ggproto("StatContourFill", ggplot2::Stat,
     required_aes = c("x", "y", "z"),
     default_aes = ggplot2::aes(fill = ..level_mid.., order = ..level..),
+    dropped_aes = "z",
     setup_params = function(data, params) {
         if (is.null(params$global) || isTRUE(params$global.breaks)) {
             params$breaks <- setup_breaks(data,