+++ NEWS.md
@@ -5,6 +5,7 @@
 -   Fixed a bug in `geom_streamline()` when plotting multiple fields on the same panel.
 -   Fixes a bug with `GetTopography()` on Windows.
 -   Updates documentation to use valid HTML5 per CRAN's new format.
+-   Fixes the "The following aesthetics were dropped during statistical transformation" warning for contours.
 
 # metR 0.12.0