+++ NEWS.md
@@ -7,6 +7,12 @@
 -   Updates documentation to use valid HTML5 per CRAN's new format.
 -   Fixes the "The following aesthetics were dropped during statistical transformation" warning for contours.
 
+## Breaking changes
+
+-   Due to the udunits2 package being orphaned, `ReadNetCDF()` not longer uses it to parse dates and times.
+    The homebrewed parser might be more limited, and the wild variety of netCDF files in the wild makes it hard to fully test.
+    Please report any issues you have!
+
 # metR 0.12.0
 
 ## New Features
@@ -377,3 +383,4 @@ There has been some changes in the interface of some functions for the sake of c
 
 -   Full (not perfect) documentation.
 -   Ready for serious testing.
+