+++ NEWS.md
@@ -1,4 +1,4 @@
-# metR (development version)
+# metR 0.13.0
 
 ## New features